+++ lib/modules/registry.ml
@@ -2,8 +2,7 @@ open Lunno_common
 
 type module_info = { exports : (string * Ast.ty) list }
 
-let table : (string, (string, module_info) Hashtbl.t) Hashtbl.t =
-  Hashtbl.create 8
+let table = Hashtbl.create 8
 
 let register ~namespace ~name info =
   let ns =